clk: qcom: smd-rpm: Fix num_rpm_clks for QM215
This variable should represent the number of the last RPM clock.
It's used to identify clk type for clk handoff in probe function.

In QM215 clocks, The last RPM clock is RPM_SMD_DIV_A_CLK2, but
`num_rpm_clks` is set to RPM_SMD_SYSMMNOC_A_CLK. Fix it.

Change-Id: Ic31903b13d5dbe526d0d8f8df44d22f2fe5dd770